Fix CCMP (WPA2) in preparation for 11n. This code didn't handle QoS frames correctly but QoS frames are required for 11n A-MPDU aggregation and 11n STAs are required to use CCMP instead of WEP or TKIP ciphers.
The QoS bit in FC0 is part of AAD (additional authentication data) but
was being masked unconditionally.
The FC1 order bit is masked to 0 in AAD if a data frame contains a QoS
control field but this code was looking for HT control fields instead.

Add an XXX comment about another bit which must be set if SPP (signaling
and payload protected) A-MSDUs are supported. Neither Linux nor FreeBSD
seem to set this bit, and we don't support SPP A-MSDUs yet so a comment
seems good enough for now.

ok deraadt mpi kettenis guenther
helpful hints from mikeb